Substituted int Prezzo with decimal Prezzo

+++ Backend-ProgettoSettimanale2/Models/Scarpa.cs
@@ -5,7 +5,7 @@
         public Guid Id { get; set; }
         public string? Marca { get; set; }
         public string? Modello { get; set; }
-        public int? Prezzo { get; set; }
+        public decimal? Prezzo { get; set; }
         public string? Descrizione { get; set; }
         public string? UrlCopertina { get; set; }
         public string? InputUrl { get; set; }
